countSignificantDecimals to util
